// database/prisma-oppulence/generated/zod/inputTypeSchemas/LeadBudgetExpenseUncheckedCreateInputSchema.ts
import type { Prisma } from '@prisma/client'

import { z } from 'zod'

export const LeadBudgetExpenseUncheckedCreateInputSchema: z.ZodType<Prisma.LeadBudgetExpenseUncheckedCreateInput> = z
    .object({
        id: z.string().cuid().optional(),
        budgetId: z.string(),
        amount: z.number(),
        category: z.string(),
        description: z.string(),
        receipt: z.string().optional().nullable(),
        expenseDate: z.coerce.date(),
        createdAt: z.coerce.date().optional()
    })
    .strict()

export default LeadBudgetExpenseUncheckedCreateInputSchema